Widen the namespace id from smallint to int
The 16-bit ceiling was reachable in a way the model never intended: ids are
never reclaimed, so the space counts every namespace ever created, and the burn
fixed one commit ago drained it at worker-restart rate. Widening it is a
breaking schema change and the release candidate is the last place it can land,
because JobContext.NamespaceId and NamespaceListItem are public and additive-only
from 1.0.

Nine columns move (namespaces.id and the eight namespace_id copies); the other
Int16 columns - max_attempts, failure_count, attempt_number - are unrelated and
stay. On the wire that is 115 CLR sites across 72 files, 74 provider SQL
declarations, and the two namespace-id dialect bindings; every SQL site was read
before it was touched, because nothing compiles them.

Both stamp constants move to baseline-1.0.1. They must: every M001 statement is
existence-guarded and the preflight verifies history rather than schema, so an
amended M001 under an unchanged stamp lets a smallint database boot green and
fail later on a conversion. With the stamp moved it refuses at bootstrap and
names the reprovision.

NamespaceIdBudget keeps its guard and loses its premise. At 32 bits the suite
cannot spend the space by running, so the message no longer reads as a chore
reminder: if it fires, something is allocating ids without creating namespaces,
and a reset hides that rather than fixing it.

+++ anvil/Anvil.Burst/BurstDb.cs
@@ -42,7 +42,7 @@ internal sealed class BurstDb(string provider, string schema)
     private readonly string _prefix = LocalDatabase.IsSqlite(provider) ? "" : schema + ".";
 
     /// <summary>The numeric namespace id the events and alerts tables carry, by namespace name.</summary>
-    public async Task<short> NamespaceIdAsync(string namespaceName, CancellationToken ct)
+    public async Task<int> NamespaceIdAsync(string namespaceName, CancellationToken ct)
     {
         await using var connection = await OpenAsync(ct);
         await using var command = connection.CreateCommand();
@@ -67,7 +67,7 @@ internal sealed class BurstDb(string provider, string schema)
     /// backlog larger than the one that was seeded.
     /// </remarks>
     public async Task<int> CountAlertableEventsAsync(
-        short namespaceId,
+        int namespaceId,
         IReadOnlyList<string> definitionNames,
         long exclusiveLow,
         long inclusiveHigh,
@@ -123,7 +123,7 @@ internal sealed class BurstDb(string provider, string schema)
     /// Backdates every event in the namespace by <paramref name="age"/> so the projection read's safe
     /// horizon admits them all. Returns the rows aged.
     /// </summary>
-    public async Task<int> AgeEventsAsync(short namespaceId, TimeSpan age, CancellationToken ct)
+    public async Task<int> AgeEventsAsync(int namespaceId, TimeSpan age, CancellationToken ct)
     {
         await using var connection = await OpenAsync(ct);
         await using var command = connection.CreateCommand();
