+++ src/Acta.Postgres/Sql/Execution/Jobs/UpdateJobInput.routine.sql
@@ -12,7 +12,7 @@ LANGUAGE plpgsql
 AS $$
 DECLARE
     v_from_status SMALLINT;
-    v_namespace_id SMALLINT;
+    v_namespace_id INT;
     v_lineage BIGINT;
     v_definition INT;
     v_tenant INT;
