+++ src/Acta.Postgres/Sql/Execution/Schedules/TriggerScheduleNow.routine.sql
@@ -17,7 +17,7 @@ DECLARE
     v_next TIMESTAMPTZ;
     v_version INT;
     v_slot_status SMALLINT;
-    v_ns SMALLINT;
+    v_ns INT;
     v_def INT;
     v_lineage BIGINT;
     v_en INT;
