+++ src/Acta.Postgres/Sql/Execution/Signals/RaiseSignal.routine.sql
@@ -18,7 +18,7 @@ DECLARE
     v_expired BOOLEAN := FALSE;
     v_message VARCHAR;
     v_from_status SMALLINT;
-    v_namespace_id SMALLINT;
+    v_namespace_id INT;
     v_lineage_root_id BIGINT;
     v_definition_id INT;
     v_tenant_id INT;
