+++ src/Acta.Postgres/Sql/Maintenance/PurgeExpiredData.routine.sql
@@ -4,7 +4,7 @@
 DROP FUNCTION IF EXISTS {{schema}}.purge_expired_data;
 
 CREATE OR REPLACE FUNCTION {{schema}}.purge_expired_data(
-    p_namespace_id SMALLINT,
+    p_namespace_id INT,
     p_events_retention_days INT,
     p_alert_retention_days INT,
     p_worker_retention_seconds INT,
